+++ config.go
@@ -8,6 +8,7 @@ import (
 	"strings"
 
 	mptls "github.com/absmach/mproxy/pkg/tls"
+	"github.com/absmach/mproxy/pkg/tls/verifier"
 	"github.com/caarlos0/env/v10"
 )
 
@@ -22,22 +23,22 @@ func (c *Config) EnvParse(opts env.Options) error {
 	switch {
 	case len(opts.FuncMap) == 0:
 		opts.FuncMap = map[reflect.Type]env.ParserFunc{
-			reflect.TypeOf(c.TLSConfig.ClientValidation.ValidationMethods): parseSliceValidateMethod,
-			reflect.TypeOf(mptls.OCSP):                                     parseValidateMethod,
+			reflect.TypeOf(c.TLSConfig.Verifier.ValidationMethods): parseSliceValidateMethod,
+			reflect.TypeOf(verifier.OCSP):                          parseValidateMethod,
 		}
 	default:
-		opts.FuncMap[reflect.TypeOf(c.TLSConfig.ClientValidation.ValidationMethods)] = parseSliceValidateMethod
-		opts.FuncMap[reflect.TypeOf(mptls.OCSP)] = parseValidateMethod
+		opts.FuncMap[reflect.TypeOf(c.TLSConfig.Verifier.ValidationMethods)] = parseSliceValidateMethod
+		opts.FuncMap[reflect.TypeOf(verifier.OCSP)] = parseValidateMethod
 	}
 	return env.ParseWithOptions(c, opts)
 }
 
 func parseSliceValidateMethod(v string) (interface{}, error) {
-	var vms []mptls.ValidateMethod
+	var vms []verifier.ValidateMethod
 	v = strings.TrimSpace(v)
 	vmss := strings.Split(v, ",")
 	for _, vm := range vmss {
-		v, err := mptls.ParseValidateMethod(vm)
+		v, err := verifier.ParseValidateMethod(vm)
 		if err != nil {
 			return nil, err
 		}
@@ -47,5 +48,5 @@ func parseSliceValidateMethod(v string) (interface{}, error) {
 }
 
 func parseValidateMethod(v string) (interface{}, error) {
-	return mptls.ParseValidateMethod(v)
+	return verifier.ParseValidateMethod(v)
 }
